m1

+++ faster_rcnn_codes/fast_rcnn/test.py
@@ -138,7 +138,7 @@ def im_detect(net, im, boxes=None):
         im_blob = blobs['data']
         blobs['im_info'] = np.array(
             [[im_blob.shape[2], im_blob.shape[3], im_scales[0]]],
-            dtype=np.float16)
+            dtype=np.float32)
 
     #print(im_blob.shape)
     #print(blobs['im_info'].shape)
